requirements.txt
discord.py>=2.3.0
Flask>=2.3.0
numpy
scipy
soundfile
PyYAML
torch==2.0.1